combined flows tests into single files
