+++ test/e2e/specs/flows/setup-onboarding__flows.spec.ts
@@ -1,0 +1,194 @@
+import {
+	BrowserManager,
+	NewTestUserDetails,
+	NewUserResponse,
+	NewSiteResponse,
+	RestAPIClient,
+} from '@automattic/calypso-e2e';
+import { tags, test, expect } from '../../lib/pw-base';
+import { apiCloseAccount, apiDeleteSite } from '../shared';
+
+test.describe(
+	'Setup Onboarding Flows',
+	{
+		tag: [ tags.CALYPSO_RELEASE ],
+	},
+	() => {
+		const accountsToCleanup: {
+			testUser: NewTestUserDetails;
+			newUserDetails: NewUserResponse;
+			newSiteDetails?: NewSiteResponse;
+		}[] = [];
+
+		test( 'As a new user, I can complete the onboarding flow and purchase a domain with a plan', async ( {
+			page,
+			componentDomainSearch,
+			helperData,
+			pageCartCheckout,
+			pageSignupPickPlan,
+			pageUserSignUp,
+		} ) => {
+			const planName = 'Personal';
+			const testUser = helperData.getNewTestUser();
+			let selectedDomain: string;
+			let newUserDetails: NewUserResponse;
+			let newSiteDetails: NewSiteResponse;
+
+			await test.step( 'When I enter the onboarding flow', async function () {
+				BrowserManager.setStoreCookie( page, { currency: 'USD' } );
+				await page.goto( helperData.getCalypsoURL( '/setup/onboarding' ) );
+			} );
+
+			await test.step( 'And I sign up as a new user', async function () {
+				newUserDetails = await pageUserSignUp.signupSocialFirstWithEmail( testUser.email );
+			} );
+
+			await test.step( 'And I search for a domain', async function () {
+				await componentDomainSearch.search( helperData.getBlogName() );
+			} );
+
+			await test.step( 'And I add the first suggestion to the cart', async function () {
+				selectedDomain = await componentDomainSearch.selectFirstSuggestion();
+			} );
+
+			await test.step( 'And I continue to the next step', async function () {
+				await componentDomainSearch.continue();
+			} );
+
+			await test.step( `And I select the ${ planName } plan`, async function () {
+				newSiteDetails = await pageSignupPickPlan.selectPlan( planName );
+				accountsToCleanup.push( { testUser, newUserDetails, newSiteDetails } );
+			} );
+
+			await test.step( 'Then I see the plan at checkout', async function () {
+				await pageCartCheckout.validateCartItem( `WordPress.com ${ planName }` );
+			} );
+
+			await test.step( 'And I see the domain at checkout', async function () {
+				await pageCartCheckout.validateCartItem( selectedDomain );
+			} );
+		} );
+
+		test( 'As a new user, I can complete the onboarding flow and purchase a plan without a custom domain', async ( {
+			page,
+			componentDomainSearch,
+			helperData,
+			pageCartCheckout,
+			pageSignupPickPlan,
+			pageUserSignUp,
+		} ) => {
+			const testUser = helperData.getNewTestUser();
+			const planName = 'Personal';
+			let newUserDetails: NewUserResponse;
+			let newSiteDetails: NewSiteResponse;
+
+			await test.step( 'When I enter the onboarding flow', async function () {
+				BrowserManager.setStoreCookie( page, { currency: 'USD' } );
+				await page.goto( helperData.getCalypsoURL( '/setup/onboarding' ) );
+			} );
+
+			await test.step( 'And I sign up as a new user', async function () {
+				newUserDetails = await pageUserSignUp.signupSocialFirstWithEmail( testUser.email );
+			} );
+
+			await test.step( 'And I skip domain purchase', async function () {
+				await componentDomainSearch.search( helperData.getBlogName() );
+				await componentDomainSearch.skipPurchase();
+			} );
+
+			await test.step( `And I select the ${ planName } plan`, async function () {
+				newSiteDetails = await pageSignupPickPlan.selectPlan( planName );
+				accountsToCleanup.push( { testUser, newUserDetails, newSiteDetails } );
+			} );
+
+			await test.step( 'Then I see the plan at checkout', async function () {
+				await pageCartCheckout.validateCartItem( `WordPress.com ${ planName }` );
+			} );
+
+			await test.step( 'And I see only one item in the cart', async function () {
+				await pageCartCheckout.validateCartItemsCount( 1 );
+			} );
+		} );
+
+		test( 'As a new user, I can complete the onboarding flow with domain transfer and plan purchase', async ( {
+			page,
+			componentDomainSearch,
+			helperData,
+			pageCartCheckout,
+			pageSignupPickPlan,
+			pageUseADomainIAlreadyOwn,
+			pageUserSignUp,
+		} ) => {
+			const testUser = helperData.getNewTestUser();
+			const blogName = helperData.getBlogName();
+			const domain = 'a8ctesting.com';
+			const planName = 'Personal';
+			let newUserDetails: NewUserResponse;
+			let newSiteDetails: NewSiteResponse;
+
+			await test.step( 'When I enter the onboarding flow', async function () {
+				await page.goto( helperData.getCalypsoURL( '/setup/onboarding' ) );
+			} );
+
+			await test.step( 'And I sign up as a new user', async function () {
+				newUserDetails = await pageUserSignUp.signupSocialFirstWithEmail( testUser.email );
+			} );
+
+			await test.step( 'And I search for a domain', async function () {
+				await componentDomainSearch.search( blogName );
+			} );
+
+			await test.step( 'And I click the "Use a domain I already own" button', async function () {
+				await componentDomainSearch.clickUseADomainIAlreadyOwn();
+			} );
+
+			await test.step( 'And I fill the "Use a domain I own" input', async function () {
+				expect( await pageUseADomainIAlreadyOwn.getDomainInputValue() ).toBe( blogName );
+				await pageUseADomainIAlreadyOwn.fillUseDomainIOwnInput( domain );
+			} );
+
+			await test.step( 'And I select the "Transfer your domain" option', async function () {
+				await pageUseADomainIAlreadyOwn.selectTransferYourDomain();
+			} );
+
+			await test.step( `And I select the ${ planName } plan`, async function () {
+				newSiteDetails = await pageSignupPickPlan.selectPlan( planName );
+				accountsToCleanup.push( { testUser, newUserDetails, newSiteDetails } );
+			} );
+
+			await test.step( 'Then I see the plan at checkout', async function () {
+				await pageCartCheckout.validateCartItem( `WordPress.com ${ planName }` );
+			} );
+
+			await test.step( 'And I see the domain at checkout', async function () {
+				await pageCartCheckout.validateCartItem( domain );
+			} );
+		} );
+
+		test.afterAll( 'Delete all user accounts generated', async function () {
+			for ( const account of accountsToCleanup ) {
+				const restAPIClient = new RestAPIClient(
+					{
+						username: account.testUser.username,
+						password: account.testUser.password,
+					},
+					account.newUserDetails.body.bearer_token
+				);
+
+				if ( account.newSiteDetails ) {
+					await apiDeleteSite( restAPIClient, {
+						url: account.newSiteDetails.blog_details.url,
+						id: account.newSiteDetails.blog_details.blogid,
+						name: account.newSiteDetails.blog_details.blogname,
+					} );
+				}
+
+				await apiCloseAccount( restAPIClient, {
+					userID: account.newUserDetails.body.user_id,
+					username: account.newUserDetails.body.username,
+					email: account.testUser.email,
+				} );
+			}
+		} );
+	}
+);
