--- test/e2e/specs/flows/start-domain__purchase-domain-and-plan.spec.ts
@@ -1,101 +1,0 @@
-import {
-	BrowserManager,
-	NewTestUserDetails,
-	NewUserResponse,
-	NewSiteResponse,
-	RestAPIClient,
-} from '@automattic/calypso-e2e';
-import { tags, test } from '../../lib/pw-base';
-import { apiCloseAccount, apiDeleteSite } from '../shared';
-
-test.describe(
-	'Domain-only flow: Purchase domain and plan',
-	{
-		tag: [ tags.CALYPSO_RELEASE ],
-	},
-	() => {
-		const planName = 'Personal';
-		let selectedDomain: string;
-		let testUser: NewTestUserDetails;
-		let newUserDetails: NewUserResponse;
-		let newSiteDetails: NewSiteResponse;
-
-		test( 'As a new user, I can complete the domain-only flow and purchase a domain with a plan', async ( {
-			page,
-			componentDomainSearch,
-			helperData,
-			pageCartCheckout,
-			pageSignupPickPlan,
-			pageUserSignUp,
-		} ) => {
-			testUser = helperData.getNewTestUser();
-
-			await test.step( 'When I enter the domain-only flow', async function () {
-				BrowserManager.setStoreCookie( page, { currency: 'USD' } );
-				await page.goto( helperData.getCalypsoURL( '/start/domain' ) );
-			} );
-
-			await test.step( 'And I search for a domain', async function () {
-				await componentDomainSearch.search( helperData.getBlogName() );
-			} );
-
-			await test.step( 'And I add the first suggestion to the cart', async function () {
-				selectedDomain = await componentDomainSearch.selectFirstSuggestion();
-			} );
-
-			await test.step( 'And I continue to the next step', async function () {
-				await componentDomainSearch.continue();
-			} );
-
-			await test.step( 'And I select "New site" option', async function () {
-				await componentDomainSearch.selectNewSite();
-			} );
-
-			await test.step( `And I select the ${ planName } plan`, async function () {
-				await pageSignupPickPlan.selectPlanWithoutSiteCreation(
-					planName,
-					new RegExp( '.*start/domain/user-social.*' )
-				);
-			} );
-
-			await test.step( 'And I sign up as a new user and wait for site creation', async function () {
-				[ newUserDetails, newSiteDetails ] =
-					await pageUserSignUp.signupWithEmailAndWaitForSiteCreation( testUser.email );
-			} );
-
-			await test.step( 'Then I see the plan at checkout', async function () {
-				await pageCartCheckout.validateCartItem( `WordPress.com ${ planName }` );
-			} );
-
-			await test.step( 'And I see the domain at checkout', async function () {
-				await pageCartCheckout.validateCartItem( selectedDomain );
-			} );
-		} );
-
-		test.afterEach( 'Delete site and user accounts generated', async function () {
-			if ( newUserDetails && testUser ) {
-				const restAPIClient = new RestAPIClient(
-					{
-						username: testUser.username,
-						password: testUser.password,
-					},
-					newUserDetails.body.bearer_token
-				);
-
-				if ( newSiteDetails ) {
-					await apiDeleteSite( restAPIClient, {
-						url: newSiteDetails.blog_details.url,
-						id: newSiteDetails.blog_details.blogid,
-						name: newSiteDetails.blog_details.blogname,
-					} );
-				}
-
-				await apiCloseAccount( restAPIClient, {
-					userID: newUserDetails.body.user_id,
-					username: newUserDetails.body.username,
-					email: testUser.email,
-				} );
-			}
-		} );
-	}
-);
