+++ test/e2e/specs/flows/start-domain__purchase-flows.spec.ts
@@ -1,0 +1,148 @@
+import {
+	BrowserManager,
+	NewTestUserDetails,
+	NewUserResponse,
+	NewSiteResponse,
+	RestAPIClient,
+} from '@automattic/calypso-e2e';
+import { tags, test } from '../../lib/pw-base';
+import { apiCloseAccount, apiDeleteSite } from '../shared';
+
+test.describe(
+	'Start Domain Only Flows',
+	{
+		tag: [ tags.CALYPSO_RELEASE ],
+	},
+	() => {
+		const accountsToCleanup: {
+			testUser: NewTestUserDetails;
+			newUserDetails: NewUserResponse;
+			newSiteDetails?: NewSiteResponse;
+		}[] = [];
+
+		test( 'As a new user, I can complete the domain-only flow and purchase a domain with a plan', async ( {
+			page,
+			componentDomainSearch,
+			helperData,
+			pageCartCheckout,
+			pageSignupPickPlan,
+			pageUserSignUp,
+		} ) => {
+			const testUser = helperData.getNewTestUser();
+			const planName = 'Personal';
+			let selectedDomain: string;
+			let newUserDetails: NewUserResponse;
+			let newSiteDetails: NewSiteResponse;
+
+			await test.step( 'When I enter the domain-only flow', async function () {
+				BrowserManager.setStoreCookie( page, { currency: 'USD' } );
+				await page.goto( helperData.getCalypsoURL( '/start/domain' ) );
+			} );
+
+			await test.step( 'And I search for a domain', async function () {
+				await componentDomainSearch.search( helperData.getBlogName() );
+			} );
+
+			await test.step( 'And I add the first suggestion to the cart', async function () {
+				selectedDomain = await componentDomainSearch.selectFirstSuggestion();
+			} );
+
+			await test.step( 'And I continue to the next step', async function () {
+				await componentDomainSearch.continue();
+			} );
+
+			await test.step( 'And I select "New site" option', async function () {
+				await componentDomainSearch.selectNewSite();
+			} );
+
+			await test.step( `And I select the ${ planName } plan`, async function () {
+				await pageSignupPickPlan.selectPlanWithoutSiteCreation(
+					planName,
+					new RegExp( '.*start/domain/user-social.*' )
+				);
+			} );
+
+			await test.step( 'And I sign up as a new user and wait for site creation', async function () {
+				[ newUserDetails, newSiteDetails ] =
+					await pageUserSignUp.signupWithEmailAndWaitForSiteCreation( testUser.email );
+				accountsToCleanup.push( { testUser, newUserDetails, newSiteDetails } );
+			} );
+
+			await test.step( 'Then I see the plan at checkout', async function () {
+				await pageCartCheckout.validateCartItem( `WordPress.com ${ planName }` );
+			} );
+
+			await test.step( 'And I see the domain at checkout', async function () {
+				await pageCartCheckout.validateCartItem( selectedDomain );
+			} );
+		} );
+
+		test( 'As a new user, I can complete the domain-only flow and purchase just a domain (no site)', async ( {
+			page,
+			componentDomainSearch,
+			helperData,
+			pageCartCheckout,
+			pageUserSignUp,
+		} ) => {
+			const testUser = helperData.getNewTestUser();
+			let selectedDomain: string;
+			let newUserDetails: NewUserResponse;
+
+			await test.step( 'When I enter the domain-only flow', async function () {
+				BrowserManager.setStoreCookie( page, { currency: 'USD' } );
+				await page.goto( helperData.getCalypsoURL( '/start/domain' ) );
+			} );
+
+			await test.step( 'And I search for a domain', async function () {
+				await componentDomainSearch.search( helperData.getBlogName() );
+			} );
+
+			await test.step( 'And I add the first suggestion to the cart', async function () {
+				selectedDomain = await componentDomainSearch.selectFirstSuggestion();
+			} );
+
+			await test.step( 'And I continue to the next step', async function () {
+				await componentDomainSearch.continue();
+			} );
+
+			await test.step( 'And I select "Just buy a domain" option', async function () {
+				await componentDomainSearch.selectJustBuyADomain();
+			} );
+
+			await test.step( 'And I sign up as a new user', async function () {
+				newUserDetails = await pageUserSignUp.signupWithEmail( testUser.email );
+				accountsToCleanup.push( { testUser, newUserDetails } );
+			} );
+
+			await test.step( 'Then I see the domain at checkout', async function () {
+				await pageCartCheckout.validateCartItem( selectedDomain );
+			} );
+		} );
+
+		test.afterAll( 'Delete all user accounts generated', async function () {
+			for ( const account of accountsToCleanup ) {
+				const restAPIClient = new RestAPIClient(
+					{
+						username: account.testUser.username,
+						password: account.testUser.password,
+					},
+					account.newUserDetails.body.bearer_token
+				);
+
+				if ( account.newSiteDetails ) {
+					await apiDeleteSite( restAPIClient, {
+						url: account.newSiteDetails.blog_details.url,
+						id: account.newSiteDetails.blog_details.blogid,
+						name: account.newSiteDetails.blog_details.blogname,
+					} );
+				}
+
+				await apiCloseAccount( restAPIClient, {
+					userID: account.newUserDetails.body.user_id,
+					username: account.newUserDetails.body.username,
+					email: account.testUser.email,
+				} );
+			}
+		} );
+	}
+);
